requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
python-docx>=0.8.11
Pillow>=9.0.0
pyautogui>=0.9.50
pandas>=2.0.0 